Update all OPC.pde files to same version
We were running different versions of the OPC.pde fil which led to
inconsistencies in how we should initialize the display. Now all the
processing examples and skeleton code are using the same version of the
file. We are using the most updated version from the official
repository and not the one from the 2.0 release, because the 2.0
version caused errors with PApplets that the most current version does
not.